Moved draw methods into GPUImageProgram

+++ framework/Source/GPUImagePosterizeFilter.h
@@ -3,6 +3,6 @@
 @interface GPUImagePosterizeFilter : GPUImageFilter
 
 // The number of color levels to reduce the image space to. This ranges from 1 to 256, with a default of 10.
-@property(readwrite, nonatomic) NSUInteger colorLevels; 
+@property(readwrite, nonatomic) GLfloat colorLevels; 
 
 @end
